+++ renderer/src/components/Sync-old-backup/Dropzone/Dropzone.js
