+++ renderer/src/components/Sync-old-backup/Gitea/Gitea.js
@@ -1,28 +1,14 @@
-import { useEffect, useState } from 'react';
+import React from 'react';
 import {
   AuthenticationContextProvider,
   RepositoryContextProvider,
 } from 'gitea-react-toolkit';
 import GiteaFileBrowser from './GiteaFileBrowser';
 import { environment } from '../../../../environment';
-import { createSyncProfile } from '../Ag/SyncToGiteaUtils';
 
-const Gitea = ({ setAuth, setRepo }) => {
-  const [authentication, setAuthentication] = useState();
-  const [repository, setRepository] = useState();
-
-  useEffect(() => {
-    setAuth(authentication);
-    setRepo(repository);
-    // on auth change update sycn on user profile
-    (async () => {
-      if (authentication !== undefined) {
-        await createSyncProfile(authentication);
-      }
-    })();
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [authentication, repository]);
-
+const Gitea = () => {
+  const [authentication, setAuthentication] = React.useState();
+  const [repository, setRepository] = React.useState();
   return (
     <AuthenticationContextProvider
       config={{
