+++ renderer/src/components/Sync-old-backup/ProgressBar.js
