+++ renderer/src/components/Sync/Gitea/EditorSync/EditorSyncUtils.js
