+++ renderer/src/components/Sync/Gitea/Gitea.js
@@ -1,14 +1,28 @@
-import React from 'react';
+import { useEffect, useState } from 'react';
 import {
   AuthenticationContextProvider,
   RepositoryContextProvider,
 } from 'gitea-react-toolkit';
 import GiteaFileBrowser from './GiteaFileBrowser';
 import { environment } from '../../../../environment';
+import { createSyncProfile } from '../Ag/SyncToGiteaUtils';
 
-const Gitea = () => {
-  const [authentication, setAuthentication] = React.useState();
-  const [repository, setRepository] = React.useState();
+const Gitea = ({ setAuth, setRepo }) => {
+  const [authentication, setAuthentication] = useState();
+  const [repository, setRepository] = useState();
+
+  useEffect(() => {
+    setAuth(authentication);
+    setRepo(repository);
+    // on auth change update sycn on user profile
+    (async () => {
+      if (authentication !== undefined) {
+        await createSyncProfile(authentication);
+      }
+    })();
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [authentication, repository]);
+
   return (
     <AuthenticationContextProvider
       config={{
