+++ renderer/src/components/Sync/Gitea/GiteaFileBrowser.js
@@ -1,95 +1,44 @@
-import React, { useContext } from 'react';
+import React, { useContext, useEffect } from 'react';
 import {
   AuthenticationContext,
   RepositoryContext,
-  createContent, readContent, get,
+  get,
 } from 'gitea-react-toolkit';
 import { useTranslation } from 'react-i18next';
 import { ChevronRightIcon } from '@heroicons/react/24/solid';
-import localforage, * as localForage from 'localforage';
-import { SnackBar } from '@/components/SnackBar';
-import { validate } from '@/util/validate';
-import { checkDuplicate } from '@/core/burrito/importBurrito';
-import moment from 'moment';
-import { AutographaContext } from '@/components/context/AutographaContext';
 import { SyncContext } from '../SyncContextProvider';
 import * as logger from '../../../logger';
-import Dropzone from '../Dropzone/Dropzone';
-import {
-  importServerProject, createSyncProfile, handleCreateRepo, createFiletoServer, updateFiletoServer,
-} from './GiteaUtils';
-import ProgressBar from '../ProgressBar';
-import ProjectMergePop from './ProjectMergePop';
 import LoadingSpinner from '../LoadingSpinner';
+import GridRow from '../GridRow';
+import ProjectMergePop from './ProjectMerge/ProjectMergePopUp';
 
 /* eslint-disable no-console */
 // eslint-disable-next-line react/prop-types
 const GiteaFileBrowser = ({ changeRepo }) => {
   const {
-    states: { dragFromAg },
+    states: { selectedGiteaProject, syncProgress, refreshGiteaListUI },
     action: {
-      setDragFromAg, handleDropToAg, setTotalUploadedAg, settotalFilesAg, setUploadstartAg,
-      fetchProjects,
+      setSelectedGiteaProject,
     },
-  } = React.useContext(SyncContext);
-
-  const {
-    action: {
-      setNotifications,
-      // setActiveNotificationCount,
-    },
-  } = useContext(AutographaContext);
-
-  const [updateGitea, setUpdateGitea] = React.useState(false);
-  const [refreshGiteaListUI, setRefreshGiteaListUI] = React.useState({ triger: false, timeOut: false });
+  } = useContext(SyncContext);
 
   const { t } = useTranslation();
-  const [totalUploaded, setTotalUploaded] = React.useState(0);
-  const [uploadStart, setUploadstart] = React.useState(false);
-  // const [uploadDone, setUploadDone] = React.useState(false);
-  const [totalFiles, settotalFiles] = React.useState(0);
-  const [merge, setMerge] = React.useState(false);
 
   // eslint-disable-next-line no-unused-vars
   const [advacnedOption, setAdvacnedOption] = React.useState(false);
   const [projects, setProjects] = React.useState([]);
   const [files, setFiles] = React.useState([]);
+  // active step is used in adavacned sync option future
   // eslint-disable-next-line no-unused-vars
   const [activeStep, setActiveStep] = React.useState();
 
-  const [projectObj, setProjectObj] = React.useState({});
-  const [sbData, setSbData] = React.useState({});
-  const [userBranch, setUserBranch] = React.useState({});
   const [steps, setSteps] = React.useState([]);
-  const [snackBar, setOpenSnackBar] = React.useState(false);
-  const [snackText, setSnackText] = React.useState('');
-  const [notify, setNotify] = React.useState();
-  const [model, setModel] = React.useState({
-    openModel: false,
-    title: '',
-    confirmMessage: '',
-    buttonName: '',
-    });
   const { state: auth, component: authComponent } = useContext(
     AuthenticationContext,
   );
   const { state: repo, component: repoComponent } = useContext(
     RepositoryContext,
   );
-
-  const addNewNotification = async (title, text, type) => {
-    localforage.getItem('notification').then((value) => {
-      const temp = [...value];
-      temp.push({
-          title,
-          text,
-          type,
-          time: moment().format(),
-          hidden: true,
-      });
-      setNotifications(temp);
-    });
-  };
 
   const fetchTree = async (treeUrl, projectName) => {
     logger.debug('Dropzone.js', 'calling fetchTree event');
@@ -116,320 +65,6 @@ const GiteaFileBrowser = ({ changeRepo }) => {
     setProjects(tree);
     setFiles(blob);
   };
-  const getPath = (filename) => {
-    logger.debug('Dropzone.js', 'calling getPath event');
-    const arr = [];
-    if (steps.length > 1) {
-      steps.forEach((s) => {
-        arr.push(s.name);
-      });
-      arr.shift();
-    }
-    arr.push(filename);
-    logger.debug('Dropzone.js', 'returning path with filename');
-    return (arr.join('/'));
-  };
-  const readData = async (value) => {
-    logger.debug('Dropzone.js', 'calling readData event');
-    const filePath = getPath(value.path);
-    await readContent(
-      {
-        config: auth.config,
-        owner: auth.user.login,
-        repo: repo.name,
-        ref: 'master',
-        filepath: filePath,
-      },
-    ).then((result) => {
-      logger.debug('Dropzone.js', 'sending the data from Gitea with content');
-      handleDropToAg({ result: { ...result, from: 'gitea' } });
-    });
-  };
-
-  const modelClose = () => {
-    if (!model.buttonName === 'Replace') {
-      settotalFiles(0);
-      handleDropToAg(null);
-    }
-    setModel({
-      openModel: false,
-      title: '',
-      confirmMessage: '',
-      buttonName: '',
-    });
-  };
-
-  const callImport = async (updateBurrito) => {
-    modelClose();
-    logger.debug('Dropzone.js', 'Inside callImport');
-    // console.log('inside IMport function , all validation success : ');
-    await importServerProject(
-      updateBurrito,
-      repo,
-      sbData,
-      auth,
-      userBranch,
-        {
-          setUploadstart,
-          setTotalUploaded,
-        },
-    ).finally(async () => {
-          setUploadstart(false);
-          setTotalUploaded(0);
-          settotalFiles(0);
-          handleDropToAg(null);
-          // refresh left pane local
-          await fetchProjects();
-          await addNewNotification(
-            'Sync',
-            'Project Sync to Ag successfull',
-            'success',
-          );
-        });
-    };
-
-  // Read gitea folder structure and data after drag (Gitea - AG )
-  const readGiteaFolderData = async (repo, from) => {
-    logger.debug('Dropzone.js', 'calling read Gitea Folder Data event');
-    // console.log("read folder data : ", from);
-    // get all branch details
-    if (from === 'gitea') {
-      console.log('fetching branches', from);
-      await fetch(`https://git.door43.org/api/v1/repos/${repo.owner.username}/${repo.name}/branches`)
-        .then((resposne) => resposne.json())
-        .then((branchData) => {
-          localForage.getItem('userProfile').then(async (user) => {
-            const regex = new RegExp(`${user.username }/\\d{4}-\\d{2}-\\d{2}.1`);
-            const userProjectBranch = branchData.find((value) => regex.test(value.name));
-            // const userBranch = 'mater';
-            if (userProjectBranch) {
-              setUserBranch(userProjectBranch);
-              await readContent(
-                {
-                  config: auth.config,
-                  owner: auth.user.login,
-                  repo: repo.name,
-                  ref: userProjectBranch?.name,
-                  filepath: 'metadata.json',
-                },
-              ).then(async (result) => {
-                logger.debug('Dropzone.js', 'sending the data from Gitea with content');
-                await fetch(result.download_url)
-                  .then((resposne) => resposne.json())
-                  .then(async (metaContent) => {
-                  // console.log('gitea meta : ', metaContent);
-                  // Validate Burrito (check for metadata.json)
-                  const sb = Buffer.from(metaContent.data);
-                  const metaDataSb = JSON.parse(sb);
-                  setSbData(metaDataSb);
-                  const success = await validate('metadata', 'gitea/metadata.json', sb, metaDataSb.meta.version);
-                  console.log('success : ', success);
-                  if (success) {
-                    // get total file count to fetch
-                    logger.debug('GiteaFileBrowser.js', 'import project from remote , validate burrito success');
-                    settotalFiles(Object.keys(metaDataSb?.ingredients).length);
-                    // check project exist
-                    const duplicate = await checkDuplicate(metaDataSb, user.username, 'projects');
-                    console.log('duplicate : ', duplicate);
-                    if (duplicate) {
-                      logger.warn('GiteaFileBrowser.js', 'Project already available');
-                      setProjectObj({
-                        repo,
-                        userProjectBranch,
-                        metaDataSbRemote: metaDataSb,
-                        agUsername: user?.username,
-                        auth,
-                      });
-                      duplicate && setMerge(true);
-                    } else {
-                      callImport(false);
-                    }
-                  } else {
-                    logger.debug('GiteaFileBrowser.js', 'Burrito Validation Failed');
-                    console.log('Burrito Validation Failed');
-                    setNotify('failure');
-                    setSnackText('Burrito Validation Failed');
-                    setOpenSnackBar(true);
-                    await addNewNotification(
-                      'Sync',
-                      'Project Sync to Ag Failed - Burrito Validation Failed',
-                      'failure',
-                    );
-                  }
-                  });
-              }).catch(async (err) => {
-                logger.debug('GiteaFileBrowser.js', 'Invalid Project , Burrito not found', err);
-                // console.log(" Error burrito not found", err);
-                setNotify('failure');
-                setSnackText('Invalid Project , Burrito not found');
-                setOpenSnackBar(true);
-                await addNewNotification(
-                  'Sync',
-                  'Project Sync to Ag Failed - Invalid Project , Burrito not found',
-                  'failure',
-                );
-              });
-          } else {
-            logger.debug('Dropzone.js', 'Invalid Project , No Valid Branch');
-            // console.log("Invalid Project directory no valid Branch Found");
-            setNotify('failure');
-            setSnackText('Invalid Project directory no valid Branch Found');
-            setOpenSnackBar(true);
-            await addNewNotification(
-              'Sync',
-              'Project Sync to Ag Failed - Invalid Project directory no valid Branch Found',
-              'failure',
-            );
-          }
-          });
-        }).catch(async (error) => {
-            setNotify('failure');
-            setSnackText('Something Went Wrong .. check internet', error);
-            setOpenSnackBar(true);
-            await addNewNotification(
-              'Sync',
-              `Project Sync to Ag Failed - Something Went Wrong .. check internet'- ${error}`,
-              'failure',
-            );
-            logger.debug('GiteaFileBrowser.js', 'Something Went Wrong .. check internet', error);
-        });
-    } else {
-      handleDropToAg(null);
-      setDragFromAg(null);
-    }
-  };
-
-  // Drag from Ag - Gitea
-  const handleDropFolder = async (data) => {
-    logger.debug('Dropzone.js', 'calling handleDropFolder event');
-    if (data?.result?.from === 'autographa') {
-      logger.debug('Dropzone.js', 'data send from Autographa');
-      const projectData = data.result.projectMeta[0];
-      const projectId = Object.keys(projectData.identification.primary.ag)[0];
-      const projectName = projectData.identification.name.en;
-      const ingredientsObj = projectData.ingredients;
-      const projectCreated = projectData.meta.dateCreated.split('T')[0];
-      const repoName = `ag-${projectData.languages[0].tag}-${projectData.type.flavorType.flavor.name}-${projectName.replace(/[\s+ -]/g, '_')}`;
-
-      localForage.getItem('userProfile').then(async (user) => {
-        const newpath = localStorage.getItem('userPath');
-        const fs = window.require('fs');
-        const path = require('path');
-        const projectsMetaPath = path.join(newpath, 'autographa', 'users', user?.username, 'projects', `${projectName}_${projectId}`);
-        settotalFilesAg(Object.keys(ingredientsObj).length);
-        // Create A REPO for the project
-        await handleCreateRepo(repoName.toLowerCase(), auth).then(
-          async (result) => {
-          if (result.id) {
-            setUploadstartAg(true);
-            setUpdateGitea(true);
-            // Successfully created , upload new files to repo
-            // console.log("inside success creation", result.name);
-            console.log('start uploading');
-            logger.debug('Dropzone.js', 'calling handleDropFolder event - Repo created : syncing started');
-            // read metadata
-            // const Metadata = fs.readFileSync(path.join(projectsMetaPath, 'metadata.json'), 'utf8');
-            const Metadata = fs.readFileSync(path.join(projectsMetaPath, 'metadata.json'));
-            await createFiletoServer(JSON.stringify(Metadata), 'metadata.json', `${user?.username}/${projectCreated}.1`, result.name, auth);
-            // Read ingredients
-            /* eslint-disable no-await-in-loop */
-            /* eslint-disable no-restricted-syntax */
-            for (const key in ingredientsObj) {
-              if (Object.prototype.hasOwnProperty.call(ingredientsObj, key)) {
-                setTotalUploadedAg((prev) => prev + 1);
-                const Metadata1 = fs.readFileSync(path.join(projectsMetaPath, key), 'utf8');
-                await createContent({
-                  config: auth.config,
-                  owner: auth.user.login,
-                  repo: result.name,
-                  branch: `${user?.username}/${projectCreated}.1`,
-                  filepath: key,
-                  content: Metadata1,
-                  message: `commit ${key}`,
-                  author: {
-                    email: auth?.user?.email,
-                    username: auth?.user?.username,
-                  },
-                // eslint-disable-next-line no-unused-vars
-                }).then((res) => {
-                  logger.debug('Dropzone.js', `file uploaded to Gitea ${key}`);
-                  // console.log('RESPONSE :', res);
-                })
-                // eslint-disable-next-line no-unused-vars
-                .catch((err) => {
-                  logger.debug('Dropzone.js', `failed to upload file to Gitea ${key}`);
-                  setUploadstartAg(false);
-                  setUpdateGitea(false);
-                  // console.log(key, ' : error : ', err);
-                });
-            }
-          }
-            setDragFromAg(null);
-            setUploadstartAg(false);
-            setUpdateGitea(false);
-            settotalFilesAg(0);
-            setTotalUploadedAg(0);
-            handleDropToAg(null);
-            logger.debug('Dropzone.js', 'calling handleDropFolder event - syncing Finished');
-            console.log('Finish uploading');
-            await addNewNotification(
-              'Sync',
-              'Project Sync to Gitea Success',
-              'success',
-            );
-            }
-          },
-
-          async (error) => {
-            // error creation , conflict already exist 409, update content if there.
-            logger.debug('Dropzone.js', 'calling handleDropFolder event - Repo already exist : ', error.message);
-            logger.debug('Dropzone.js', 'calling handleDropFolder event - started update project : ');
-            if (error.message.includes('409')) {
-              console.log('started update project ');
-              setUploadstartAg(true);
-              setUpdateGitea(true);
-              const metadataContent = fs.readFileSync(path.join(projectsMetaPath, 'metadata.json'));
-              await updateFiletoServer(JSON.stringify(metadataContent), 'metadata.json', `${user?.username}/${projectCreated}.1`, repoName, auth);
-              // Read ingredients and update
-              for (const key in ingredientsObj) {
-                if (Object.prototype.hasOwnProperty.call(ingredientsObj, key)) {
-                  const metadata1 = fs.readFileSync(path.join(projectsMetaPath, key), 'utf8');
-                  setTotalUploadedAg((prev) => prev + 1);
-                  await updateFiletoServer(metadata1, key, `${user?.username}/${projectCreated}.1`, repoName, auth);
-              }
-            }
-              setDragFromAg(null);
-              setUploadstartAg(false);
-              setUpdateGitea(false);
-              settotalFilesAg(0);
-              setTotalUploadedAg(0);
-              logger.debug('Dropzone.js', 'calling handleDropFolder event - update syncing Finished');
-              console.log('Finish updating project');
-              await addNewNotification(
-                'Sync',
-                'Project Sync to Gitea Success',
-                'success',
-              );
-            } else {
-              logger.debug('Dropzone.js', 'calling handleDropFolder event - Repo Updation Error : ', error.message);
-              setUploadstartAg(false);
-              setUpdateGitea(false);
-              await addNewNotification(
-                'Sync',
-                `Project Sync to Gitea falied - ${error.message}`,
-                'failure',
-              );
-            }
-          },
-          );
-          setUploadstartAg(false);
-          setUpdateGitea(false);
-      });
-    } else {
-      handleDropToAg(null);
-      setDragFromAg(null);
-    }
-  };
 
   const handleStep = (step) => () => {
     logger.debug('Dropzone.js', 'calling handleStep event');
@@ -445,7 +80,8 @@ const GiteaFileBrowser = ({ changeRepo }) => {
     setSteps([]);
     changeRepo();
   };
-  React.useEffect(() => {
+
+  useEffect(() => {
     if (files.length === 0 && projects.length === 0) {
       if (repo) {
         fetchTree(`${auth.config.server}/${repo?.tree_url}`, repo.name);
@@ -454,32 +90,20 @@ const GiteaFileBrowser = ({ changeRepo }) => {
   // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [repo?.tree_url]);
 
-  // const testingButtonClick = async () => {
-  //   console.log('clicked');
-  // };
-
-  React.useEffect(() => {
-    (async () => {
-      if (auth !== undefined) {
-        await createSyncProfile(auth);
-      }
-    })();
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [auth]);
-
-  React.useEffect(() => {
-    if (updateGitea && !refreshGiteaListUI.triger) {
-      setRefreshGiteaListUI({ ...refreshGiteaListUI, triger: true });
-    }
-    if (!updateGitea && refreshGiteaListUI.triger) {
-      setRefreshGiteaListUI({ ...refreshGiteaListUI, timeOut: true });
-      setTimeout(() => {
-        setRefreshGiteaListUI({ ...refreshGiteaListUI, triger: false, timeOut: false });
-      }, 500);
-    }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [updateGitea]);
-
+  // Auto refresh for gitea side after upload from Ag move to context
+  // React.useEffect(() => {
+  //   if (updateGitea && !refreshGiteaListUI.triger) {
+  //     setRefreshGiteaListUI({ ...refreshGiteaListUI, triger: true });
+  //   }
+  //   if (!updateGitea && refreshGiteaListUI.triger) {
+  //     setRefreshGiteaListUI({ ...refreshGiteaListUI, timeOut: true });
+  //     setTimeout(() => {
+  //       setRefreshGiteaListUI({ ...refreshGiteaListUI, triger: false, timeOut: false });
+  //     }, 500);
+  //   }
+  // // eslint-disable-next-line react-hooks/exhaustive-deps
+  // }, [updateGitea]);
+  console.log({ selectedGiteaProject });
   return (
     (!auth && authComponent)
     || (!repo && (
@@ -487,11 +111,8 @@ const GiteaFileBrowser = ({ changeRepo }) => {
         <div>
           {refreshGiteaListUI.timeOut === true ? <LoadingSpinner /> : <div>{repoComponent}</div>}
         </div>
-        <div className="row-span-6">
-          <Dropzone dropped={() => handleDropFolder(dragFromAg)} />
-        </div>
       </div>
-))
+  ))
     || (
     <>
       <div className="flex flex-row mx-5 my-3 border-b-1 border-primary">
@@ -503,7 +124,7 @@ const GiteaFileBrowser = ({ changeRepo }) => {
         {steps.map((label, index) => (
           (steps.length - 1 === index)
             ? (
-              <span key={label} className="font-semibold tracking-wide text-primary " onClick={handleStep(label)} role="button" tabIndex={-1}>
+              <span key={label} className="font-semibold tracking-wide text-primary">
                 <ChevronRightIcon className="h-4 w-4 mx-2 inline-block fill-current text-gray-800" aria-hidden="true" />
                 {label.name}
               </span>
@@ -515,138 +136,26 @@ const GiteaFileBrowser = ({ changeRepo }) => {
             )
         ))}
       </div>
-      {uploadStart
-        && <ProgressBar currentValue={totalUploaded} totalValue={totalFiles} />}
 
-      {merge
-        && (
+      {selectedGiteaProject?.mergeStatus && (
         <ProjectMergePop
-          setMerge={setMerge}
-          projectObj={projectObj}
-          addNewNotification={addNewNotification}
-        />
-)}
+          selectedGiteaProject={selectedGiteaProject}
+          setSelectedGiteaProject={setSelectedGiteaProject}
 
+        />
+      )}
+
+      {/* advanced and some of the functions are for future use - current sycn is project wise and future is file wise */}
       {!advacnedOption
         && (
-        <table className="min-w-full divide-y divide-gray-200" data-testid="table">
-          <thead className="bg-gray-50">
-            <tr>
-              <th
-                scope="col"
-                className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider"
-                data-testid="th-name"
-              >
-                {t('label-name')}
-              </th>
-            </tr>
-          </thead>
-          <tbody className="bg-white divide-y divide-gray-200">
-            {/* <tr key={repo.name} draggable onDragStart={() => readGiteaFolderData(repo)}> */}
-            <tr key={repo.name} draggable onDragStart={() => handleDropToAg({ result: { repo, from: 'gitea', readGiteaFolderData } })}>
-              <td className="px-6 py-4 whitespace-nowrap cursor-pointer">
-                <div className="flex items-center">
-                  {/* <svg viewBox="0 0 14 16" fill="none" className="mr-2 h-5 w-5 text-gray-500" xmlns="http://www.w3.org/2000/svg">
-                    <path fillRule="evenodd" clipRule="evenodd" d="M9.09182 1.00903H3.95833C2.3696 1.00903 1 2.29684 1 3.88635V12.1526C1 13.8316 2.28009 15.1703 3.95833 15.1703H10.1227C11.7122 15.1703 13 13.7428 13 12.1526V5.08002L9.09182 1.00903Z" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" />
-                    <path d="M8.88916 1V3.2446C8.88916 4.34028 9.77573 5.22917 10.8706 5.23148C11.8868 5.2338 12.9262 5.23457 12.9964 5.22994" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" />
-                    <path opacity="0.4" d="M8.74248 10.8824H4.57812" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" />
-                    <path opacity="0.4" d="M7.16739 7.06128H4.57788" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" />
-                  </svg> */}
-                  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 128 128" width="25px" height="25px">
-                    <path fill="#D8D4EA" d="M100.2,104H19c-5.5,0-10-4.5-10-10V29h98.7c6,0,10.7,5.3,9.9,11.2l-7.5,55C109.5,100.2,105.2,104,100.2,104 z" />
-                    <path fill="#FFF" d="M104,104H19c-5.5,0-10-4.5-10-10V24h24.6c3.3,0,6.5,1.7,8.3,4.5l4.1,6.1c1.9,2.8,5,4.5,8.3,4.5H89 c5.5,0,10,4.5,10,10v41v1.9C99,96.5,100.8,100.8,104,104L104,104z" />
-                    <path fill="#454B54" d="M100.2,107H19c-7.2,0-13-5.8-13-13V24c0-1.7,1.3-3,3-3h24.6c4.4,0,8.4,2.2,10.8,5.8l4.1,6.1 c1.3,2,3.5,3.1,5.8,3.1H89c7.2,0,13,5.8,13,13v35c0,1.7-1.3,3-3,3s-3-1.3-3-3V49c0-3.9-3.1-7-7-7H54.4c-4.4,0-8.4-2.2-10.8-5.8 l-4.1-6.1c-1.3-2-3.5-3.1-5.8-3.1H12v67c0,3.9,3.1,7,7,7h81.2c3.5,0,6.5-2.6,6.9-6.1l7.5-55c0.2-2-0.4-4-1.7-5.5 c-1.3-1.5-3.2-2.4-5.2-2.4c-1.7,0-3-1.3-3-3s1.3-3,3-3c3.7,0,7.3,1.6,9.7,4.4c2.5,2.8,3.6,6.5,3.2,10.2l-7.5,55 C112.3,102.1,106.7,107,100.2,107z" />
-                    <path fill="#454B54" d="M107.7,32H43c-1.7,0-3-1.3-3-3s1.3-3,3-3h64.7c1.7,0,3,1.3,3,3S109.3,32,107.7,32z" />
-                  </svg>
-                  <span className="text-sm text-gray-900">
-                    {/* {console.log(repo.name.split('-').pop().replaceAll('_', " "))} */}
-                    {repo?.name}
-                    {' '}
-                    <span className="font-bold">
-                      {' '}
-                      (
-                      {repo.name.split('-').pop().replaceAll('_', ' ')}
-                      )
-                      {' '}
-                    </span>
-                  </span>
-                </div>
-              </td>
-            </tr>
-          </tbody>
-        </table>
-)}
-
-      {advacnedOption
-      && (
-      <table className="min-w-full divide-y divide-gray-200">
-        <thead className="bg-gray-50">
-          <tr>
-            <th
-              scope="col"
-              className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider"
-            >
-              {t('label-name')}
-            </th>
-            {/* <th
-                scope="col"
-                className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider"
-              >
-                Created
-              </th>
-              <th
-                scope="col"
-                className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider"
-              >
-                Last Viewed
-              </th> */}
-          </tr>
-        </thead>
-        <tbody className="bg-white divide-y divide-gray-200">
-          {projects.map((project) => (
-            <tr key={project.name} onClick={() => fetchTree(project.url, project.path)}>
-              {/* <tr key={project.name}> */}
-              {/* {console.log("project : ", projects)} */}
-              <td
-                className="px-6 py-4 whitespace-nowrap"
-              >
-                <div className="flex items-center">
-                  <div className="ml-0">
-                    <div className="text-sm text-gray-900">{project.path}</div>
-                  </div>
-                </div>
-              </td>
-              {/* <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">{file.created}</td>
-                <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">{file.updated}</td> */}
-            </tr>
-                  ))}
-        </tbody>
-        <tbody className="bg-white divide-y divide-gray-200">
-          {files.map((file) => (
-            <tr key={file.name} draggable onDragStart={() => readData(file)}>
-
-              <td className="px-6 py-4 whitespace-nowrap">
-                <div className="flex items-center">
-                  <div className="ml-0">
-                    <div className="text-sm text-gray-900">{file.path}</div>
-                  </div>
-                </div>
-              </td>
-              {/* <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">{file.created}</td>
-                <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">{file.updated}</td> */}
-            </tr>
-                  ))}
-        </tbody>
-      </table>
-)}
-      {/* <Dropzone dropped={() => handleDropFolder(dragFromAg)} /> */}
-      <SnackBar
-        openSnackBar={snackBar}
-        snackText={snackText}
-        setOpenSnackBar={setOpenSnackBar}
-        setSnackText={setSnackText}
-        error={notify}
-      />
+          <GridRow
+            title={`${repo?.name} (${repo.name.split('-').pop().replaceAll('_', ' ')})`}
+            lastSync={undefined}
+            selected
+            isUpload={selectedGiteaProject?.repo?.name === repo?.name && syncProgress.syncStarted}
+            uploadPercentage={(syncProgress.completedFiles * 100) / syncProgress.totalFiles}
+          />
+      )}
     </>
     )
   );
