+++ renderer/src/components/Sync/Gitea/ProjectMerge/ProjectMergePopUp.js
@@ -4,9 +4,9 @@ import { useTranslation } from 'react-i18next';
 import ConfirmationModal from '@/layouts/editor/ConfirmationModal';
 import PropTypes from 'prop-types';
 import PopUpModal from '@/layouts/Sync/PopUpModal';
-import { VerticalLinearStepper } from '@/components/Sync/VerticalStepperProgress';
 import { SnackBar } from '@/components/SnackBar';
 import useAddNotification from '@/components/hooks/useAddNotification';
+import { VerticalLinearStepper } from '../../VerticalStepperProgress';
 import { mergeProgressSteps } from './ProjectMergeConst';
 import LoadingSpinner from '../../LoadingSpinner';
 import { tryMergeProjects } from './MergeActions';
