+++ renderer/src/components/Sync/GridRow.js
